+++ backend/src/casual/casual.ts
@@ -1,25 +1,30 @@
+import type { CasualConfig } from "$types/data";
 import Settings from "$src/settings/settings";
 import { mapBits } from "$shared/maps";
 import { join } from "node:path";
 import fsp from "node:fs/promises";
 import Server from "$src/net/server";
 import { Message, Recieves } from "$types/messages";
-import type { CasualConfig } from "$types/data";
 import { exists } from "$src/util";
+import { watch, type FSWatcher } from "chokidar";
+import Log from "$src/log";
 
 export default class Casual {
 	static config: CasualConfig;
+	static manualSelectionChange: number[] | null = null;
 
 	static saveConfig() {
 		Settings.set("casual", this.config);
 	}
 
 	static async init() {
-		this.updateConfig();
-		Settings.on("tfPath", () => this.updateConfig());
+		this.loadConfig();
+		this.watchCasualCriteria();
+		Settings.on("tfPath", () => this.watchCasualCriteria());
 
 		Server.onConnect("casual", (reply) => {
 			reply(Message.CasualConfig, this.config);
+			if(this.manualSelectionChange) reply(Message.CasualSelectionChanged, true);
 		});
 
 		Server.on(Recieves.UpdateCasualProfile, (data, { ws }) => {
@@ -76,16 +81,59 @@ export default class Casual {
 			this.setCasualCriteria(this.config.profiles[index]!.selection);
 			this.saveConfig();
 		});
+
+		Server.on(Recieves.HandleCasualChanged, (data) => {
+			if(!this.manualSelectionChange) return;
+
+			if(data.action === "new") {
+				const id = crypto.randomUUID();
+
+				this.config.profiles.push({
+					name: data.name,
+					id,
+					selection: this.manualSelectionChange
+				});
+
+				this.config.selectedProfile = id;
+			} else if(data.action === "overwrite") {
+				const profile = this.config.profiles.find(p => p.id === data.id);
+				if(profile) {
+					profile.selection = this.manualSelectionChange;
+					this.config.selectedProfile = profile.id;
+				}
+			} else {
+				const profile = this.config.profiles.find(p => p.id === this.config.selectedProfile);
+				if(profile) this.setCasualCriteria(profile.selection);
+			}
+
+			this.manualSelectionChange = null;
+			Server.send("casual", Message.CasualConfig, this.config);
+			Server.send("casual", Message.CasualSelectionChanged, false);
+		});
 	}
-	
-	static async updateConfig() {
-		const savedConfig = Settings.get("casual");
-		if(savedConfig) {
-			this.config = savedConfig;
-		} else {
-			const selection = await this.readCasualCriteria();
-			if(!selection) return;
-	
+
+	static watcher?: FSWatcher;
+	static async watchCasualCriteria() {
+		this.watcher?.close();
+		
+		const path = this.getCriteriaPath();
+		if(!path) return;
+		
+		this.watcher = watch(path, {
+			persistent: false,
+			ignoreInitial: true
+		});
+
+		this.watcher.on("change", () => this.checkMatch());
+		this.checkMatch();
+	}
+
+	static async checkMatch() {
+		const selection = await this.readCasualCriteria();
+		if(!selection) return;
+
+		// Create a default profile picture if the user doesn't have one
+		if(!this.config) {
 			const id = crypto.randomUUID();
 			const defaultConfig: CasualConfig = {
 				profiles: [{ name: "Default Profile", id, selection }],
@@ -94,7 +142,22 @@ export default class Casual {
 
 			Settings.set("casual", defaultConfig);
 			this.config = defaultConfig;
+			return;
 		}
+
+		// If the user has manually changed their selection note it so once they go to /casual we can ask what to do
+		const profile = this.config.profiles.find(p => p.id === this.config.selectedProfile);
+		if(!profile || this.selectionsMatch(profile.selection, selection)) return;
+		if(this.manualSelectionChange && this.selectionsMatch(this.manualSelectionChange, selection)) return;
+
+		Log.info("Casual criteria changed externally");
+		this.manualSelectionChange = selection;
+		Server.send("casual", Message.CasualSelectionChanged, true);
+	}
+	
+	static loadConfig() {
+		const savedConfig = Settings.get("casual");
+		if(savedConfig) this.config = savedConfig;
 	}
 
 	static getCriteriaPath() {
@@ -129,5 +192,11 @@ export default class Casual {
 		const path = this.getCriteriaPath();
 		const text = numbers.map((n) => `selected_maps_bits: ${n}`).join("\n");
 		if(path) await fsp.writeFile(path, text);
+	}
+
+	static selectionsMatch(a: number[], b: number[]) {
+		const aMasked = a.map((bits, i) => bits & Number(mapBits[i]));
+		const bMasked = b.map((bits, i) => bits & Number(mapBits[i]));
+		return aMasked.every((bits, i) => bits === bMasked[i]);
 	}
 }