+++ .github/release-notes.md
@@ -1,2 +1,3 @@
 * Added option to enable time alive column in game 
 * tfview/respawned.cfg is automatically `exec`'d when respawning
+* Added dialog for when the user changes their casual selection externally